feat(plugins): implement new affiliation and BCL plugins with transaction handling, sync services, and event listeners for enhanced functionality
