+++ src/plugins/bcl/listeners/bcl-tx.listener.ts
@@ -2,25 +2,25 @@ import { BasePluginTxListener } from '@/mdw-sync/plugins/base-plugin-tx.listener
 import { Plugin } from '@/mdw-sync/plugins/plugin.interface';
 import { BasePluginSyncService } from '@/mdw-sync/plugins/base-plugin-sync.service';
 import { Injectable, Logger } from '@nestjs/common';
-import { DexPlugin } from '../dex.plugin';
-import { DexSyncTransactionService } from '../services/dex-sync-transaction.service';
+import { BclPlugin } from '../bcl.plugin';
+import { BclSyncTransactionService } from '../services/bcl-sync-transaction.service';
 
 @Injectable()
-export class DexTxListener extends BasePluginTxListener {
-  protected readonly logger = new Logger(DexTxListener.name);
+export class BclTxListener extends BasePluginTxListener {
+  protected readonly logger = new Logger(BclTxListener.name);
 
   constructor(
-    private readonly dexPlugin: DexPlugin,
-    private readonly dexSyncService: DexSyncTransactionService,
+    private readonly bclPlugin: BclPlugin,
+    private readonly bclSyncService: BclSyncTransactionService,
   ) {
     super();
   }
 
   protected getPlugin(): Plugin {
-    return this.dexPlugin;
+    return this.bclPlugin;
   }
 
   protected getSyncService(): BasePluginSyncService {
-    return this.dexSyncService;
+    return this.bclSyncService;
   }
 }
