+++ src/plugins/social/social-plugin.module.ts
@@ -4,13 +4,20 @@ import { SocialPlugin } from './social.plugin';
 import { Post } from '@/plugins/social/entities/post.entity';
 import { Topic } from '@/plugins/social/entities/topic.entity';
 import { Account } from '@/plugins/account/entities/account.entity';
+import { Tx } from '@/mdw-sync/entities/tx.entity';
+import { PluginSyncState } from '@/mdw-sync/entities/plugin-sync-state.entity';
 import { MDW_PLUGIN } from '@/mdw-sync/plugins/plugin.tokens';
 import { AeModule } from '@/ae/ae.module';
 import { AccountModule } from '@/plugins/account/account.module';
 import { TopicsController } from './controllers/topics.controller';
 import { PostsController } from './controllers/posts.controller';
+import { SocialTxListener } from './listeners/social-tx.listener';
+import { SocialSyncTransactionService } from './services/social-sync-transaction.service';
+
 @Module({
-  imports: [TypeOrmModule.forFeature([Post, Topic, Account])],
+  imports: [
+    TypeOrmModule.forFeature([Tx, PluginSyncState, Post, Topic, Account]),
+  ],
   providers: [
     AeModule,
     AccountModule,
@@ -19,6 +26,8 @@ import { PostsController } from './controllers/posts.controller';
       provide: MDW_PLUGIN,
       useClass: SocialPlugin,
     },
+    SocialTxListener,
+    SocialSyncTransactionService,
   ],
   exports: [SocialPlugin, TypeOrmModule],
   controllers: [PostsController, TopicsController],
